Add AVPlayerProvider
